src/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1
python-multipart==0.0.6
pydantic-settings==2.1.0
